+++ CHANGELOG.md
@@ -11,6 +11,12 @@
 - X10D: Added `Color.GetClosestConsoleColor()`
 - X10D: Added `DateTime.GetIso8601WeekOfYear()` and `DateTimeOffset.GetIso8601WeekOfYear()`
 - X10D: Added `DirectoryInfo.Clear([bool])`
+- X10D: Added `IEnumerable<T>.CountWhereNot(Func<T, bool>)`
+- X10D: Added `IEnumerable<T>.FirstWhereNot(Func<T, bool>)`
+- X10D: Added `IEnumerable<T>.FirstWhereNotOrDefault(Func<T, bool>)`
+- X10D: Added `IEnumerable<T>.LastWhereNot(Func<T, bool>)`
+- X10D: Added `IEnumerable<T>.LastWhereNotOrDefault(Func<T, bool>)`
+- X10D: Added `IEnumerable<T>.WhereNot(Func<T, bool>)`
 - X10D: Added `IList<T>.RemoveRange(Range)`
 - X10D: Added `IList<T>.Swap(IList<T>)` (#62)
 - X10D: Added `Point.IsOnLine(LineF)`, `Point.IsOnLine(PointF, PointF)`, and `Point.IsOnLine(Vector2, Vector2)`